Add a new Makefile target, 'slug', which will update the management resource group name in the config.yaml, setting it to the tre_id concatenated with the current epoch seconds, expressed in hex. 'make slug' should be run after the config.yaml has been set up, but before any other targets.
This allows creating a unique management resource group name easily, so that
deployments can be torn down and new ones deployed for the same tre_id
without having the problem of soft-deleted resources such as the Key Vault.
Just run 'make slug' after destroying the management group and before
re-deploying.

To go with this, the config.sample.yaml is updated to note that the
mgmt_resource_group_name is set automatically, the json schema is
modified to reflect the schema of the tre_id + slug, and there's a new
script, devops/scripts/generate-unique-slug.sh

The generate-unique-slug.sh script will check if the mgmt_resource_group_name
has been set already in the config file, and if it has, it will check if the
group actually exists in Azure. If it does, it aborts, to prevent the
possibility of 'zombie' resources that get forgotten after a new slug is
generated.